@@ -1,18 +1,33 @@
+import { Link, NavLink } from 'react-router-dom';
 import './appHeader.scss';
 
 const AppHeader = () => {
     return (
         <header className="app__header">
             <h1 className="app__title">
-                <a href="##">
+                <Link to="/">
                     <span>Marvel</span> information portal
-                </a>
+                </Link>
             </h1>
             <nav className="app__menu">
                 <ul>
-                    <li><a href="##">Characters</a></li>
+                    <li>
+                        <NavLink 
+                            end 
+                            style={({isActive}) => ({color: isActive ? '#9F0013' : 'inherit'})}
+                            to="/">
+                            Characters
+                        </NavLink>
+                    </li>
                     /
-                    <li><a href="##">Comics</a></li>
+                    <li>
+                        <NavLink 
+                            end 
+                            style={({isActive}) => ({color: isActive ? '#9F0013' : 'inherit'})}
+                            to="/comics">
+                            Comics
+                        </NavLink>
+                    </li>
                 </ul>
             </nav>
         </header>
